Update babel packages to match major, fix dependency issues
