+++ .gitignore
@@ -22,3 +22,4 @@ public/bundle/bundle.js.map
 public/bundle/Lekton-*
 public/bundle/style.css
 
+uploads
