+++ nginx_example_config
@@ -11,7 +11,7 @@ server {
 
   ssl_session_cache shared:SSL:20m;
   ssl_session_timeout 1440m;
-  ssl_protocols TLSv1 TLSv1.1 TLSv1.2;
+  ssl_protocols TLSv1.1 TLSv1.2;
   ssl_prefer_server_ciphers on;
   # Using list of ciphers from https://github.com/cloudflare/sslconfig
   ssl_ciphers EECDH+CHACHA20:EECDH+AES128:RSA+AES128:EECDH+AES256:RSA+AES256:EECDH+3DES:RSA+3DES:!MD5;
